requirements.txt
# Data processing
pandas==1.5.0
numpy==1.22.0
json==2.0.9
# scikit-learn==1.0.0
textblob==0.15.3
emoji==2.0.0
contractions==0.1.73
nltk==3.7.0
spacy==3.4.0
symspellpy==6.7.0

# Deep learning
torch==2.0.0
transformers==4.25.0
datasets==2.10.0
sentence-transformers==2.2.2
peft==0.3.0
bitsandbytes>=0.37.1

# MLOps tools
mlflow==2.3.0
dvc[s3]==2.10.0
dagshub==0.2.0
torchvision
# pyyaml==6.0

# HPO
optuna==3.0.0
# AWS integration
boto3==1.26.0
awscli==1.27.0

# Experiment tracking
huggingface_hub==0.12.0

# Environment management
python-dotenv==0.21.0

# Visualization
matplotlib==3.6.0
seaborn==0.12.0
plotly==5.10.0

# Notebook dependencies
ipywidgets==8.0.0
jupyter==1.0.0

# Frontend (if running locally)
streamlit==1.19.0  # Optional for quick demos